requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
astroid==2.15.5
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
countryinfo==0.1.2
cryptography==41.0.1
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.6
docutils==0.20.1
emoji-country-flag==1.3.2
exceptiongroup==1.1.1
fastapi==0.98.0
frozenlist==1.3.3
future==0.18.3
geocoder==1.38.1
geoip2==4.7.0
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
ip2geotools==0.1.6
IP2Location==8.10.0
isort==5.12.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==24.2.0
lazy-object-proxy==1.9.0
lxml==4.9.2
markdown-it-py==3.0.0
maxminddb==2.3.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
outcome==1.2.0
packaging==23.1
pip-review==1.3.0
pkginfo==1.9.6
platformdirs==3.8.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pylint==2.17.4
pyparsing==3.1.0
pyquery==2.0.0
PySocks==1.7.1
ratelim==0.1.6
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
SecretStorage==3.3.3
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
tomlkit==0.11.8
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
twine==4.0.2
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
webencodings==0.5.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0